Fk23C: Add overrides for UNIF headered games not found in database
With the recent changes to Fk23C.h, a NES 2.0 header is very much near
necessary to set variables not found on iNES 1.0 header like submappers
and the availability of a switchable CHR-RAM/CHR-ROM in some.

UNIF boards do not support submapper in header but most of the romsets
in Fk23C/Fk23CA i found are already in the database.

This PR covers only the few roms which are not in database.